Fixed a bunch of connections, and also made the Extra Node Properties works again
svn path=/trunk/KDE/kdeedu/rocs/; revision=1221510

+++ src/Core/Data.cpp
@@ -184,10 +184,10 @@ void Data::setY(int y) {
     }
 }
 
-void Data::setWidth(qreal w) {
-    d->_width = w;
+void Data::setWidth(double w) {
+    d->_width = (qreal)w;
     if (! d->_changing) {
-	emit changed();
+        emit changed();
     }
 }
